Split up the code, cleanup and comment
